+++ contexts/index.ts
@@ -1,1 +1,1 @@
-export { ChosenItemContext } from "./ChosenItemContext";
+export { ChosenItemContext } from "./ChosenContext";
